requirements.txt
--find-links https://data.dgl.ai/wheels/repo.html
--find-links https://download.pytorch.org/whl/torch/
--find-links https://download.pytorch.org/whl/torchvision/
torch == 1.11.0+cu113  
torchvision == 0.12.0+cu113 
dgl-cu113 == 0.8.1
dglgo == 0.0.1
numpy == 1.22.3
transformers == 4.18.0
pytorch_pretrained_bert == 0.6.2
tqdm
matplotlib
sklearn
Pillow
opencv-python
pandas